+++ addons/spag_factions/XEH_postInit.sqf
