+++ addons/spag_gear/Cfgs.hpp
@@ -1,0 +1,2 @@
+#include "cfgs/CfgWeapons.hpp"
+#include "cfgs/CfgVehicles.hpp"
