+++ addons/spag_gear/cfgs/CfgWeapons.hpp
