+++ addons/spag_gear/cfgs/Hats_Vehicles.hpp
