+++ addons/spag_gear/cfgs/Hats_Weapons.hpp
