+++ addons/spag_gear/cfgs/Vests_Weapons.hpp
