+++ addons/spag_main/script_mod.hpp
@@ -1,13 +1,13 @@
 // COMPONENT should be defined in the script_component.hpp and included BEFORE this hpp
-#include "script_version.hpp"
 
 #define MAINPREFIX z
 #define PREFIX spag
 
+#include "script_version.hpp"
 
-#define VERSION MAJOR.MINOR
-#define VERSION_AR MAJOR,MINOR,PATCHLVL,BUILD
-#define VERSION_STR MAJOR.MINOR.PATCHLVL.BUILD
+#define VERSION      MAJOR.MINOR
+#define VERSION_AR   MAJOR,MINOR,PATCHLVL,BUILD
+#define VERSION_STR  MAJOR.MINOR.PATCHLVL.BUILD
 
 
 #define SPAG_TAG SPAG
