+++ addons/spag_tfar/XEH_preInit.sqf
