+++ addons/spag_vehicles/cfgs/CfgBTR.hpp
