+++ addons/spag_vehicles/cfgs/CfgMI8.hpp
