+++ addons/spag_vehicles/cfgs/CfgSU25-2.hpp
