+++ addons/spag_vehicles/cfgs/CfgYak.hpp
